Clase 4

+++ src/app/components/products/products.component.ts
@@ -1,4 +1,4 @@
-import {Component, Input} from '@angular/core';
+import {Component, EventEmitter, Input, Output} from '@angular/core';
 import {zip} from 'rxjs';
 import {switchMap} from 'rxjs/operators';
 
@@ -12,24 +12,13 @@ import {ProductsService} from "../../services/products.service"
   styleUrls: ['./products.component.css']
 })
 export class ProductsComponent {
-
+  @Input() products: Product [] = [];
+  @Output() loadMore = new EventEmitter();
+  
   myShoppingCart: Product [] = [];
   total = 0;
-  @Input() products: Product [] = [];
-  @Input() limit = 10;
-  @Input()  offset = 0;
   showProductDetail = false;
-  productChosen: Product = {
-    id:'',
-    price: 0,
-    images: [],
-    title: '',
-    category: {
-      id: '',
-      name: ''
-    },
-    description: ''
-  };
+  productChosen: Product | null = null;
   statusDetail: 'loading' | 'success' | 'error' | 'init' = 'init';
 
 
@@ -37,6 +26,9 @@ export class ProductsComponent {
     this.myShoppingCart = this.storeService.getShoppingCart();
   }
 
+  onLoadMore() {
+    this.loadMore.emit();
+  }
 
   onAddToShoppingCart(product: Product) {
     this.storeService.addProduct(product);
@@ -96,23 +88,29 @@ export class ProductsComponent {
       title: 'Nuevo title',
     };
 
-    const id = this.productChosen.id;
+    const id = this.productChosen?.id;
+    if (id != null){
     this.productsService.update(id, changes)
       .subscribe(data => {
-        const productIndex = this.products.findIndex(item => item.id === this.productChosen.id);
+        const productIndex = this.products.
+          findIndex(item => item.id === this.productChosen?.id);
         this.products[productIndex] = data;
         this.productChosen = data;
       });
+    }
   }
 
   deleteProduct() {
-    const id = this.productChosen.id;
+    const id = this.productChosen?.id;
+    if (id != null){
     this.productsService.delete(id)
       .subscribe(() => {
-        const productIndex = this.products.findIndex(item => item.id === this.productChosen.id);
+        const productIndex = this.products.
+          findIndex(item => item.id === this.productChosen?.id);
         this.products.splice(productIndex, 1);
         this.showProductDetail = false;
       });
+    }
   }
 
 }
